flow_quantum_optimizer.py
#!/usr/bin/env python3
# -*- coding: utf-8 -*-

"""
FlowQuantumOptimizer.py - Module d'optimisation des gains et stratégies pour FlowGlobalDynamics™
"""

import logging
import time
import threading
from typing import Dict, Any, List, Optional

# Import local
try:
    from flow_eventbus import get_event_bus
except ImportError:
    # Support mode hors ligne
    class DummyEventBus:
        def publish(self, *args, **kwargs): pass
        def subscribe(self, *args, **kwargs): pass
        def unsubscribe(self, *args, **kwargs): pass
    
    def get_event_bus():
        return DummyEventBus()

class FlowQuantumOptimizer:
    """
    Optimiseur quantique pour les stratégies de trading et maximisation des gains
    """
    
    def __init__(self):
        """Initialisation du FlowQuantumOptimizer"""
        self.running = False
        self.thread = None
        self.logger = self._setup_logger()
        self.event_bus = get_event_bus()
        self.optimization_queue = []
        self.last_optimization = 0
        self.current_strategy = None
        
    def _setup_logger(self):
        """Configuration du logger"""
        logger = logging.getLogger("FlowQuantumOptimizer")
        logger.setLevel(logging.INFO)
        if not logger.handlers:
            handler = logging.StreamHandler()
            formatter = logging.Formatter('%(asctime)s - %(name)s - %(levelname)s - %(message)s')
            handler.setFormatter(formatter)
            logger.addHandler(handler)
        return logger
    
    def start(self):
        """Démarrage de l'optimiseur quantique"""
        if self.running:
            self.logger.warning("FlowQuantumOptimizer est déjà en cours d'exécution")
            return
            
        self.running = True
        self.thread = threading.Thread(target=self._optimization_loop, daemon=True)
        self.thread.start()
        self._register_events()
        self.logger.info("FlowQuantumOptimizer démarré")
        
    def stop(self):
        """Arrêt de l'optimiseur quantique"""
        if not self.running:
            self.logger.warning("FlowQuantumOptimizer n'est pas en cours d'exécution")
            return
            
        self.running = False
        if self.thread:
            self.thread.join(timeout=2.0)
        self._unregister_events()
        self.logger.info("FlowQuantumOptimizer arrêté")
    
    def _register_events(self):
        """Enregistrement des abonnements aux événements"""
        self.event_bus.subscribe("market_data_update", self._handle_market_data)
        self.event_bus.subscribe("risk_assessment", self._handle_risk_assessment)
        self.event_bus.subscribe("strategy_request", self._handle_strategy_request)
    
    def _unregister_events(self):
        """Désenregistrement des abonnements aux événements"""
        self.event_bus.unsubscribe("market_data_update", self._handle_market_data)
        self.event_bus.unsubscribe("risk_assessment", self._handle_risk_assessment)
        self.event_bus.unsubscribe("strategy_request", self._handle_strategy_request)
    
    def _optimization_loop(self):
        """Boucle principale d'optimisation"""
        while self.running:
            try:
                if self.optimization_queue and time.time() - self.last_optimization > 5:
                    self._perform_optimization()
                time.sleep(0.1)
            except Exception as e:
                self.logger.error(f"Erreur dans la boucle d'optimisation: {str(e)}")
    
    def _perform_optimization(self):
        """Exécute l'optimisation de stratégie"""
        try:
            if not self.optimization_queue:
                return
                
            # Simule une optimisation de stratégie
            self.logger.info("Exécution de l'optimisation quantique des stratégies...")
            time.sleep(0.5)  # Simuler le traitement
            
            # Publie les résultats optimisés
            optimized_strategy = self._generate_optimized_strategy()
            self.current_strategy = optimized_strategy
            
            self.event_bus.publish("strategy_optimized", {
                "strategy": optimized_strategy,
                "timestamp": time.time()
            })
            
            self.optimization_queue = []
            self.last_optimization = time.time()
            self.logger.info("Optimisation complétée avec succès")
        except Exception as e:
            self.logger.error(f"Erreur lors de l'optimisation: {str(e)}")
    
    def _generate_optimized_strategy(self) -> Dict[str, Any]:
        """
        Génère une stratégie optimisée basée sur les données actuelles
        
        Returns:
            Dict contenant la stratégie optimisée
        """
        # Exemple simple de stratégie
        return {
            "name": "quantum_optimized_v1",
            "risk_level": 3,
            "entry_conditions": {
                "indicators": ["ma_cross", "rsi_oversold"],
                "timeframes": ["1h", "4h"]
            },
            "exit_conditions": {
                "take_profit": 2.5,  # Pourcentage
                "stop_loss": 1.2     # Pourcentage
            },
            "position_sizing": 0.15  # Portion du capital disponible
        }
    
    def _handle_market_data(self, data: Dict[str, Any]):
        """
        Gestionnaire pour les mises à jour de données de marché
        
        Args:
            data: Données de marché
        """
        self.logger.debug(f"Données de marché reçues: {data.get('symbol', 'unknown')}")
        self.optimization_queue.append({"type": "market_data", "data": data})
    
    def _handle_risk_assessment(self, data: Dict[str, Any]):
        """
        Gestionnaire pour les évaluations de risque
        
        Args:
            data: Données d'évaluation de risque
        """
        self.logger.debug(f"Évaluation de risque reçue: niveau {data.get('risk_level', 'unknown')}")
        self.optimization_queue.append({"type": "risk", "data": data})
    
    def _handle_strategy_request(self, data: Dict[str, Any]):
        """
        Gestionnaire pour les demandes de stratégies
        
        Args:
            data: Données de demande de stratégie
        """
        self.logger.info(f"Demande de stratégie reçue pour: {data.get('context', 'general')}")
        
        # Si nous avons déjà une stratégie optimisée et qu'elle est récente, on la renvoie
        if self.current_strategy and time.time() - self.last_optimization < 60:
            self.event_bus.publish("strategy_response", {
                "strategy": self.current_strategy,
                "request_id": data.get("request_id"),
                "timestamp": time.time()
            })
        else:
            # Sinon, on lance une nouvelle optimisation
            self.optimization_queue.append({"type": "request", "data": data})

if __name__ == "__main__":
    # Test simple de l'optimiseur
    optimizer = FlowQuantumOptimizer()
    optimizer.start()
    
    # Simuler des données entrantes
    optimizer._handle_market_data({
        "symbol": "BTC/USDT",
        "price": 45000,
        "volume": 1250.5,
        "timestamp": time.time()
    })
    
    time.sleep(2)  # Attente pour traitement
    optimizer.stop()
