requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-dotenv==1.0.0
pydantic==2.3.0
python-binance==1.0.19
ccxt==3.0.79
pybit==5.0.0
pandas==2.1.0
numpy==1.25.2
loguru==0.7.0
python-multipart==0.0.6
httpx==0.24.1
python-jose==3.3.0
passlib==1.7.4
ta==0.10.2
aiohttp==3.8.5
websockets==11.0.3
